[release] First release of caravan
Isolated the caravan project and have updated it to be better
integrated into Kubernetes. Added minimal documentation and
provided a good starting location for features an integration
into Kubernetes. It has a long way to go before it is stable,
but it works well and is being used every day.